Migrate from `dist: trusty` to `dist: xenial`
Xenial LTS support was dropped in April 2019, so it's high time we move
off of it. In addition, this should give us a performance boost thanks to some infrastructure
improvements we made (mostly not starting a bunch of services when a job starts, e.g. databases).

This also requires us to update the APT syntax for chrome.