Update CI to use Xenial and some additional cleanup (#2092)
* Avoid specifying minor node version

* Migrate from `dist: trusty` to `dist: xenial`

Xenial LTS support was dropped in April 2019, so it's high time we move
off of it. In addition, this should give us a performance boost thanks to some infrastructure
improvements we made (mostly not starting a bunch of services when a job starts, e.g. databases).

This also requires us to update the APT syntax for chrome.

* Fix chrome install syntax

Apparently we were already using the newer syntax; the upgrade guide assumed we were using something older, so my last change was unnecessary/wrong

* Npm 6 is already installed on newer xenial image

* Try removing specific matrix node versions

* Add back single node_js job, see if this works